demo_system/python/requirements.txt
fastapi
uvicorn[standard]
openpyxl
python-pptx
pandas
google-generativeai
chromadb
python-dotenv
networkx
pdfplumber
python-docx
